Refactored structure to have input handling in main, modes receive string, not individual chars

+++ mylib/s4435360_hal_hamming.h
@@ -21,7 +21,7 @@ typedef struct {
 	uint8_t decodedOutput;
 	uint16_t fullDecodedOutput;
 	int uncorrectableError;
-	uint8_t errorMask;
+	uint16_t errorMask;
 
 } HammingDecodedOutput;
